requirements.txt
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.6.15
cffi==1.17.1
channels==4.3.0
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
hf-xet==1.1.4
httptools==0.6.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
openai-whisper==20240930
packaging==25.0
protobuf==6.31.1
pycparser==2.22
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
silero-vad==5.1.2
sniffio==1.3.1
soundfile==0.13.1
sox==1.5.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
